chore: Rename Constant class
